添加打包插件
