+++ grpc-java-polaris-examples/quickstart-example/consumer/src/main/java/com/tencent/polaris/grpc/HiConsumer.java
@@ -14,8 +14,8 @@
  * the License.
  */
 
-import com.tencent.polaris.grpc.HelloGrpc;
-import com.tencent.polaris.grpc.HelloPolaris;
+package com.tencent.polaris.grpc;
+
 import com.tencent.polaris.grpc.resolver.PolarisNameResolverProvider;
 import io.grpc.ManagedChannel;
 import io.grpc.ManagedChannelBuilder;
@@ -26,16 +26,18 @@ import lombok.extern.slf4j.Slf4j;
  * @author lixiaoshuang
  */
 @Slf4j
-public class HelloConsumer {
+public class HiConsumer {
     
     public static void main(String[] args) {
-        NameResolverRegistry.getDefaultRegistry().register(new PolarisNameResolverProvider());
-        ManagedChannel channel = ManagedChannelBuilder.forTarget("polaris://grpc-demo-java:8080?namespace=default")
-                .usePlaintext().build();
         
-        HelloGrpc.HelloBlockingStub helloBlockingStub = HelloGrpc.newBlockingStub(channel);
-        HelloPolaris.request request = HelloPolaris.request.newBuilder().setMsg("hello polaris").build();
-        HelloPolaris.response response = helloBlockingStub.sayHello(request);
-        log.info("response:{}", response.getData());
+        NameResolverRegistry.getDefaultRegistry().register(new PolarisNameResolverProvider());
+        ManagedChannel channel = ManagedChannelBuilder.forTarget("polaris://Hi:8080?namespace=default").usePlaintext()
+                .build();
+        
+        HiGrpc.HiBlockingStub hiBlockingStub = HiGrpc.newBlockingStub(channel);
+        HelloPolaris.request request = HelloPolaris.request.newBuilder().setMsg("hi polaris").build();
+        HelloPolaris.response response = hiBlockingStub.sayHi(request);
+        System.out.println("grpc server response ---------> :" + response.getData());
+        System.exit(1);
     }
 }
