+++ grpc-java-polaris-examples/quickstart-example/provider/src/main/java/com/tencent/polaris/grpc/HiProvider.java
@@ -14,17 +14,17 @@
  * the License.
  */
 
-import com.tencent.polaris.grpc.HelloGrpc;
-import com.tencent.polaris.grpc.HelloPolaris;
+package com.tencent.polaris.grpc;
+
 import io.grpc.stub.StreamObserver;
 
 /**
  * @author lixiaoshuang
  */
-public class HelloProvider extends HelloGrpc.HelloImplBase {
+public class HiProvider extends HiGrpc.HiImplBase {
     
     @Override
-    public void sayHello(HelloPolaris.request request, StreamObserver<HelloPolaris.response> responseObserver) {
+    public void sayHi(HelloPolaris.request request, StreamObserver<HelloPolaris.response> responseObserver) {
         String msg = request.getMsg();
         HelloPolaris.response response = HelloPolaris.response.newBuilder().setData(msg).build();
         responseObserver.onNext(response);
